+++ pom.xml
@@ -12,7 +12,7 @@
 
 	<groupId>jp.igapyon.oiyokan</groupId>
 	<artifactId>oiyokan-initializr</artifactId>
-	<version>0.1.20210504a</version>
+	<version>0.1.20210504b</version>
 	<name>oiyokan-initializr</name>
 	<description>Oyekan Initializr is a generator for creating `oiyokan-settings.json`.</description>
 	<url>https://github.com/igapyon/oiyokan-initializr</url>
@@ -87,9 +87,14 @@
 			<version>1.20</version>
 		</dependency>
 
-		<!-- v0.5 にて spring-boot-starter-thymeleaf および spring-boot-starter-web 
-			に対応したい -->
-
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-thymeleaf</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
@@ -132,7 +137,13 @@
 
 	<build>
 		<plugins>
-			<!-- TODO Web対応時に spring-boot-maven-plugin を組み込み -->
+			<plugin>
+				<groupId>org.springframework.boot</groupId>
+				<artifactId>spring-boot-maven-plugin</artifactId>
+				<configuration>
+					<mainClass>jp.oiyokan.initializr.OiyokanInitializrApp</mainClass>
+				</configuration>
+			</plugin>
 		</plugins>
 	</build>
 </project>
