+++ src/main/java/jp/oiyokan/initializr/OiyokanInitializrConstants.java
@@ -19,5 +19,5 @@ package jp.oiyokan.initializr;
  * Oiyokan Initializr の定数.
  */
 public class OiyokanInitializrConstants {
-    public static final String VERSION = "0.1.20210504a";
+    public static final String VERSION = "0.1.20210504b";
 }
