+++ src/main/resources/templates/index.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+
+<head>
+  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+  <meta charset="UTF-8" />
+  <meta name="description" content="">
+  <title>Oiyokan Initializr</title>
+</head>
+
+<body>
+  <h3>Oiyokan Initializr</h3>
+  <p>aaa.</p>
+
+  <form class="needs-validation" th:action="@{/entry}" th:object="${databaseBean}">
+    <ul>
+      <li> <input type="text" class="" id="name" th:field="*{name}" /></li>
+      <li> <input type="text" class="" id="type" th:field="*{type}" /></li>
+      <li> <input type="text" class="" id="description" th:field="*{description}" /></li>
+      <li> <input type="text" class="" id="jdbcDriver" th:field="*{jdbcDriver}" /></li>
+      <li> <input type="text" class="" id="jdbcUrl" th:field="*{jdbcUrl}" /></li>
+      <li> <input type="text" class="" id="jdbcUser" th:field="*{jdbcUser}" /></li>
+      <li> <input type="text" class="" id="jdbcPassPlain" th:field="*{jdbcPassPlain}" /></li>
+    </ul>
+
+    <button class="" type="submit" name="connTest">Connection Test</button>
+    <button class="" type="submit" name="download">ZIP Download</button>
+
+  </form>
+
+
+</body>
+
+</html>
